Fix testcase for mul_add that was being prematurely optmized and failing
Original use case of having a literal 1, instead of any other number in simplify_mul_add, resulted in the find_unit_mult_const function to optimize away the literal 1 causing this test to fail. on the final check. Switched the constant to a non zero & one value, and now correctly passes.