Add flag for execution provider in onnxruntime allow for backwards comparability
This prevents us from issues when someone uses a newer onnxruntime version
with stream sync support but such support doesn't exist for the earlier migraphx
versions. The EP should default to the old way of inferences (with hipDeviceSynchronize())
instead of the new event based stream synchronization when this flag is missing.