Add test case for stream sync
Verify that we get the correct result by running an async graph of
operations that resolve to a known value.

First test verifies the kernel created for verification is indeed valid

Second test verifies that if we run through a series of operations
correctly and we use the compare kernel to verify the final result of
STREAM_SYNC_TEST_VAL is in the buffer. Each operation (migraphx eval,
and kernel.launch() ) is done on seperate streams in the gpu

Third test verifies the same as the second test but performs both things
on the same stream.