Fix for parse_split to handle in newer split op
Newer split moves the split attribute to an input. In this case we check the
number of input args then.

This changes allows us to move the accumulte check to outside each conditional branch
Added more debug on this as well to show more details on this failure mode.

Adds an additional test case using a generated split-13 operator styled
test.