Add the ability to show infered shape for models
Allows for a model to be converted to the same opset but turn on infer_shapes through
onnx. This allows us to get an idea of what should be valid for the output of nodes
in a network.

Usecase:

python3 tools/convert_onnx_version.py --model <model_name> --opset=<same_as_model> --infer_shapes --output <new_model_name>