Update parse_splitToSequence
Per the spec. Perform split based on splits. If Split is a scalar use the value
in the split and split equally.

If not scalar, split along desired dimensions. Check dims accordingly so they
sum to the length of the target axes

For keepdims, only split down to new sequences if keepdims=0 otherwise use the
axes length as the output of keepdims=1