Refactor parse_if tests to use add_ vs insert_ instructions and fix returns
Change things in the test cases so that we're not just replicating what
the code does but use add_instruction to dictate what we should expect
for the output of fixing empty const cases.

Had to also switch an insert to add of a literal in the empty case to achieve
this in parse_if as well.

Moved the return instructions to the end of each subgraph to also fix readability
of each test.