Regen error cases to avoid inline/folding
In the case of error cases with the IF operator, during parsing having a constant
for the input condition results in us bypassing both branches entirely and
folding the output to be one of the branches. This skips any other checks
as the other branch can be optimized into a "don't care" and pruned.

For our error case tests, they need to be regenerated as they were using a constant
constant condition which was returning early. Regenerating the onnx files after
some quick changes in gen_onnx.py solves this.