Add two test cases using the given example in the onnx Gather documentation
Generates output from the given onnx description of gather found here:
https://github.com/onnx/onnx/blob/main/docs/Operators.md#gather

Using this as a baseline as I seem to pass on the zero case but fail on the
axis = 1 case. In the other tests of shape (3,3) there appears to be a similar
error when axis = -1 which after tuning is axis = 1.