Backup of trying to get gather working
Tried to get a proper templated shape of out_comp but right now this
seems to break as I can't just update the length of a shape and get a proper
output of the strides. Currently this breaks/asserts.

I think this is the cause of axis > 0 failing since we're not getting proper gathering
for the other axes as a result and get repeated rows with the wrong data.