Remove length based shape jit constructor and calculate_strides()
These work in tandem to create a shape via the calculate strides() call. Seemed
to introduce more issues than fix, since we don't have access to resize()

Right now this is cleanup but I had used rev_partial_sum and the multiplies()
template operator created in algorithm to achieve this during debugging for
gather. The idea here would be we would statically create array() with calculate_strides()
to fix the empty stride dimensions.