license_stamper: Update hasKeySequence() to limit lines searched
Since there's no point to check for a licence twice, just break after we find
the key_messag we're looking for otherwise break at some predetermined line
limit. Saves us from having to drudge through all of a file.

A license should be at the top of a file so anything more than 10 lines past
the beginning of the file (since we're also stamping .ipynb files) should be sufficient