Make ONNXRUNTIME_COMMIT env variable via reading in .onnxrt-commit
This allows us to just update the one file with the appropriate sha-1 from
the onnxrt repo without needing to break all other scripts and things.

Will push in the github action after this commit once that is ready to turn on
our CI which just updates the onnxrt-commit.txt file weekly.