Add changes to inline IF module if arguments can be evaluated during compile
Allows us to inline and remove the IF operator if we know the input is a constant
that can be evaluated as we parse in the graph.

This avoids a bunch of shape checks, and such when certain graphs use IF operators
with only one branch populated.

Had to modify onnx_parser to make sure we don't remove previous inputs during the
parsing process when things are inlined as this causes errors with out of order nodes