Adding testcase for large max boxes and sanity nms test
Realize we didn't actually have a compute test for nonmaxsuppression to
just run this during our test runs. Add it as a part of comparrison with
the case where we have a huge max for max_output_boxes_per_class, which is
similar to what we've seen from resnext50 (issue #1283)

Test just passes if the compute() function can run and not error out for now
to sanity test that we don't stall or something weird.