Change parse_graph() to return void to return std::vector<instruction_ref>
This allows us to remove the redundant remove_return instruction during
parsing and inlining while removing the need to insert a return for inlined operators
in onnx parsing. Returns the output_ins that are parsed in from the graph instead
of void.

Had to stub out some additional calls for their return values such as parse_loop
which seems like it could also be a candidate for potentially inlining as well.