Create next_box from next_top_score once during IOU suppression
We're continually creating/destroying batch box in the while() check as we
run through the boxes_heap() by calling batch_box() constantly.

Make this next_box and only calculate it before we pop that box from the boxes_heap.

should get rid of function overhead of constant calls in the case of a large
batch size