[mlir] Improve context handling to potentially solve threading bugs
Update `mlir_program` to only create one dialect registry, and to call
registerRocMLIRPasses() (which is needed and may not be thread-safe)
exactly once. This might resolve issues with parallel compiles (or it
won't, testing that is a future PR).

In addition, use a single thread pool across all contexts. This is
recommended practice upstream for libraries that perform a lot of
compile jobs, and saves on the overhead of creating and destroying a
lot of threads (which MLIR uses to, for example, process functions in
parallel).

Finally, stop setting the unregistered dialects flag, as it is not
needed.