Perhaps actually fix the issue, or at least something related
Once I'd started temporarily building with a debug build of
MLIR (which, perhaps, could've been an asserts-on build) I found that
we were sometimes passing null pointers in to the attribute append
functions.

Further debugging (since I now had an error that I could reproduce
locally) showed that, when we were trying to add an empty vector (ex.
adding attributes when there were none), there'd be a NULL data
pointer passed in to the relevant MLIR C API function, which would
then assert. I suspect that, in less careful builds, this might be
leading to funky memory reads off the edge of something or to
segfaults.

In order to resolve this issue, guard all the mlirOperationStateAdd*()
calls with a check to make sure we're adding at least one thing.

It is not currently clear if this'll solve the problem we're seeing
when we use a non-debug MLIR build, but it can't hurt.