[MLIR] Changes for tuning API v2, MLIR grid layout changes.
This PR constitutes the MIGraphX-side changes needed to not break the
build in the presence of
https://github.com/ROCmSoftwarePlatform/rocMLIR/pull/1136 , and
updates what data is sent in to MLIR during the kernel generation and
tuning process.

Specifically, as of
https://github.com/ROCmSoftwarePlatform/rocMLIR/pull/1145 (which
hasn't landed yet but is likely to) we will want to know the number of
CUs on the card we're targeting in order to efficiently tile our
matrix multiplications. This commit therefore updates the tsv
parser (since we'll be adding a new field) and updates the compilation
process to pass num_cu into MLIR.

This commit also removes a legacy "xdlopsv2" attribute that is no
longer used as we have a more robust architecture feature lookup in
MLIR and no longer require it.