[MLIR] Capture diagnostics, handle compile failure
Add mlir_logger, which registers a MLIR diagnostic handler that
captures any information generated by a MLIR compile and saves it to a
string.

This will be useful during tuning, where some such errors may be the
result of an inapplicable tuning configuration and should be
suppressed.

However, if the high-level pipeline rejects an input, this is almost
certainly an error in the MLIR construction and an exception.