Check installed card and set to ENV variable GPU_TARGET for compilation
Set the GPU_TARGET environment variable when we invoke cmake to setup the project build
before compilation. This allows us to pull the gpu from rocminfo if none is specified
as part of project setup with cmake. If no gfx target is specified or detected
on the system, cmake will throw an error and require a user to set the GPU_TARGETS
variable prior to compilation for cross-compilation.