[MLIR] Updates needed for general stride support
This is a companion PR to
https://github.com/ROCmSoftwarePlatform/rocMLIR/pull/1312 .
The updated commit hash points onto that PR branch, so coordinated
merges are advised.

With the above rocMLIR changes, the MLIR MIGraphX dialect now
represents both the dimensions and strides of tensors inside MLIR,
thus allowing NHWC convolutions to be correctly offloaded.

In this PR, we:
- Remove special handling for the case where non-standard shapes
become input to MLIR modules
- Fold broadcast and multibroadcast operations into the input size od
MLIR modules
- Update tests
- Add an extra TRACE_MLIR print to help debug crashes in the
high-level pipeline