Cleanup and add Run parameter for sample runs
Clean up how we run samples with prints around verbose output. Allow user to set the amount of runs for batched inferences when we meausure timing.

Done to allow user to average out the effect of cold start spikes in timing. Defaulted things to 100 to mirror what we do with driver perf.