Fix bug found in parse_dynamicquantizelinear
Bug with how we were calculating intermediate zero point (interm_zp) when parsing in the operator.

Worked for uint8 without any issues due to the 0 range for uint8_t::min but when we shift things for int8 [-128, 127], we break where the zero point should be. This was caught by the verification tests after a simplify_dynamicquantizelinear pass was performed as the quantized result had the incorrect shift.

-Updated the parseer tests to the correct state
-Update simplify pass to also preserve order of q_min/q_max literals in the pass.