Fix sanitizer error with int8 overflow in unit test. Remove same_type on dot()
Ran into issue with the unit tests which test after the simplify_dynamicquantizelinear() due to overflow of the
output when performing dot (y,y) where y is an int8 vector with values that overflow. Clang ASAN was (correctly) flagging this as an error.

As a workaround I've added a subtract node such that the result out of the dot is a zero. Also I still required to remove the same_types() check for the dot() operator as this throws an error during the replace call which occurs depth first to update lower nodes instead of breadth first.

Onnx test files, parse and verify tests have been updated accordingly