Handle the input to the quantizelinear as part of pass
Doing this and then splicing part of the graph back allows us to avoid issues seen when just replacing the input. We adjust the inputs with converts to the dot/conv operator that are uint8 and then splice in proper converts/new quantize linear as part of the pass

Doing this allows us to re-add back the same_type() check onto the dot() operator

I also found that the quantizelinear op was actually errornously assuming same_type() on all inputs (2 or 3) and as a result we were failing when not needed for the output type of quantizelinear. We should infact be only checking that the input data and scale are the same, and only the zero point must be identical to that of the type of the output.