Fix parsing of MatMulInteger
Originally we were using dot and converting the input for MatMulinteger() as it was parsed in.  As a result we were inserting dot() instead of quant_dot into the model.

I've added a change to check for similar data types and then add a shift to an input  where there's a mismatch of int8 and uint8 on input. This is usually seen when dynamicquantizelinear is used as the input as it inserts a uint8 as the zero point which in some models (like gpt2) is operated on.