Initial crossentropyloss operator outline
Need to handle the compute shape here as i'll return a tuple of loss_tensor, rearranged weight tensor

Compute should be easier here as we're just multiplying weights and  or assigning things accordingly to the new output tensor for the calculation.

As usual, backing this up incase the sun explodes.