Remove transpose and fix input as columns instead
Remove the transpose ops here as this was masking an issue related to input checks for the scale and bias inputs. Supposed to match input for the scale value based on the input column value instead of the row. With this in mind we can remove the transpose here.

Updated parser tests to handle this correclty. Retested and validated output since transpose changes the math here.