Fix memory_coloring pass when MIGRAPHX_NSTREAMS > 2 (#3757)
I noticed allocation segments reaching close to uint64 max value, which clearly would throw an out of memory error when trying to allocate on the GPU. This happened when MIGRAPHX_NSTREAMS was set to 2 or greater and the model somehow was large enough to trigger it.

Changing from auto to size_t seems to fix the issue.

After some further investigation, the segment start value is the culprit which becomes int32. I've updated n as size_t for clarity.