rollback output buffer naming convention (#4007)
PR #3870 changed the output buffer naming for some reason. Previously graphs with a single output would have output buffer name main:#output_0, but after this PR this name is changed to output. This causes issues in torch-migraphx and would require version checking logic to behave differently for this change. If there is no good reason to have this change, I would like to roll it back to the way it was before.

Quick note for multiple outputs, the naming for that is unchanged: main:#output_0, main:#output_1 ... main:#output_n. It's unintuitive to suddenly have different naming for the one output case