Print MigraphX consumed Env Variables (#4015)
This change is to print only the actual consumed Env variables. Note, it doesn't verify the value of the consumed variable.

This change also guards against any version changes. If an Env variable isn't supported in a particular version, the final dump of the consumed variables will not show it.