GridSample Linear Sampler Refactor (#4067)
This PR refactors the linear sampler used in the GridSample operator to use less literals. Also adds support and tests for when channels > 1.

Note: This PR does not change the nearest and bicubic samplers. Although they are currently not used in any models we support, they should likely be refactored as well sometime in the future.