Slice Channels Conv Optimization (slice output fusion) (#4037)
Works off the conv-slice-channels branch to make a matcher that reshapes inputs to a slice->convolution in NHWC such that the slice occurs on the slowest memory. The reshapes will then be fused into the previous kernel with the added find_mlir_output_reshapes such that the slicing effectively occurs as an output fusion.