Add timestamps and driver duration (#4095)
While we have the ability to view how long a model took to run with perf, we don't get a good idea of how long the entire run with the driver took (compilation + runtime + overhead). This can be annoying especially with runs that take a long time to complete, such as when exhaustive tune is enabled.

By adding both the start and end timestamps as well as the elapsed time, we can get a more accurate idea of how long certain runs take. While it may be more beneficial to add additional timestamps and timing for the compilation stage, it would require more changes to the code base (such as the addition of a logger) that will take longer to implement. This PR aims to be a temporary solution to at least provide slightly more information when using the driver.