Skip horizontal fusion when we would concat a constant across a broadcasted axis (#4123)
Currently we dont handle such rewrites with broadcasts, which can lead to larger constants in the model. In the future, we could relax this more and allow the fusions when all the dimensions are the same size or are multiples of the other dimensions. However, when they are completely unrelated like 154 and 4096, I dont see an easy to rewrite this to avoid the constants becoming larger.