Split into stages for handling inputs, attributes and stage one/two of calculation
Cleaned this up to handle things appropriatley and be more readable since we had everything in parse() as just a large flattented piece of code. In reality the two stages to layer norm can be seperated out.

This allows us to reuse a bunch of items and checks while leveraging std::tuple() to handle input/attrs