Add in stash type attr for layernorm and additional test coverage (#4147)
Gets rid of "Stash type not supported" warnings in models whenever there is layernorm present

default behavior of layernorm op is stash_type = 1 for this attribute. Without this we throw the warning for every graph. Proper way is to upconvert first input for stash and then downconvert to type for the second scale and bias stage.

https://onnx.ai/onnx/operators/onnx__LayerNormalization.html

Added additional tests to validate and ensure we're using the proper types for input, scale and bias as well as the outputs are within the proper types for this parser.

Had to update gen_onnx and generate a few more cases to ensure layernorm is being parsed in with the correct inputs before we do any sort of calculations which would propagate through the graph