Check for hipblaslt workspace size during tuning (#4182)
We used default workspace size of 256 MB during tuning, however, it turns out
that some of the solutions require a workspace size larger than 256MB.
This causes memory access failures.

To fix this, we skip over solutions that require more than 128MB of workspace size.
The reason for choosing 128MB is for efficient memory usage.