Improve the algorithm for module sort (#4176)
Previously we only used sort in our tests, but #4049 uses it in the actual pass so we need to have a more robust implementation to handle more complicated graphs.

The previous algorithm didnt check if a node was visited or not, which means in complex graphs we would visit the nodes multiple times. It also didnt handle dead code that well either.

The new algorithm now completes in linear time. Many unit tests have been added to test for correctness and check algorithm complexity. The previous implementation would hang on these test but the new implementation finishes in under 20ms for a debug build.