Graphviz output improvement (node color change) (#4214)
This commit adds coloring for nodes in order to provide information on:

- constant foldable
- alias information
- context free information

Based on whether it is an operation that can be constant folded or has a combination of alias and context free information, we color the nodes differently. This keeps the original coloring of the node intact to aid in identifying operations at a glance, with the border adding extra information when examined more closely.

This also standardizes use of hex string representation of RGB colors, instead of using name colors. For example, instead of `Crimson`, we use `#DC143C`.