Ensure we're always parsing properly typed scales
Scales inputs should always be of type float. Its an error if we attempt to use scales or downcast the input to a lower type which could lead to errors. The onnx spec specifies across all Resize/Upsamples that scale is always of type float.

Fixed this since we originally had an else case when checking type for size or scale input  that would fall through allowing half, bf16, etc types through. Added an explicit check here and a test case for this