Add python API for returning serialized bytes (#4200)
Currently torch-migraphx has to use the load-save api to obtain serialized bytes which requires read write from file system. The msgpack bytearray is used to save migraphx program as part of the MGXModule saved using the torch.save api.

This PR allows directly encoding and decoding a program to bytearray without needing to save to disk