Rewrite transpose/reshape/broadcast between pointwise/reduce operators (#3978)
This will rewrite transpose/reshape/broadcast so they will not be between pointwise and reduction using the shape_transform_descriptor. This is similiar to the `rewrite_reshapes` that is used for pointwise/reduce fusion, however this is done before fusions in simplify_reshapes. It also supports more cases such as reduce->squeeze->pointwise.

This could be expanded in the future to support other operators like argmin/argmax/concat etc.