Streamline find_matches function (#4316)
## Motivation
Refactor `find_matches` to help improve performance by moving many things outside of the loop.

## Technical Details
The `find_matches` function would do many things inside of the loop even when tracing is not enabled:

- Checking if the trace filter matches
- Run time on the apply function
- Call to `matcher()` to construct the matcher, which is usually a no-op but in some cases we are constructing maps, etc, so we can benefit from constructing it once

This also separates the checking with and without tracing enabled, so the tracing wont interfere with optimizations the compiler might do.